feat(a card component): a dinamic card component
a dinamic card component with props
