+++ src/App.js
@@ -23,7 +23,7 @@ const GlobalStyle = createGlobalStyle`
   margin: 0;
   box-sizing: border-box;
   outline: none;
-  font-family: monospace;
+  font-family: Open Sans;
 }
 
 body {
